fix(docker)!: root-less docker container
The docker container now runs under the `retrom` user again.

BREAKING CHANGE: The UID:GID of the retrom user in the container
has changed from 1505:1505 to the more common 1000:1000. This is so
it is less likely one needs to change the ID's at runtime. You may
need to adjust or remove any `user: ` directives in your docker compose
file.