fix(docker)!: create DB dir
The directory for the embedded database in the retrom
Docker container has changed from `/app/data` to `/app/data/db`.

For those using the embedded DB ( i.e. not using an external PostgreSQL
DB ), you may notice that your data is missing. Follow the
[migration guide](https://github.com/JMBeresford/retrom/wiki/Migration-Guides#v070)
in order to retain your existing data

BREAKING CHANGE: refer to the migration guide for details