chore: ensure workspace deps are synced after build
When using `injectWorkspacePackages` w/ pnpm, the hardlinked packages
are not automatically synced. This mean that after running a build
for package A, any package B that depends on A will have an empty
or partial module to work with. This change ensures that such
packages are synced whenever a build command is completed.