feat: serve web client directly from service (#356)
The web client is now served directly by the main Retrom service. This means that the web client can now be accessed under the same port that the service listens on. This also sets the Retrom service up for distribution methods other than docker. Keep your eyes open for native binaries for Windows / MacOS / Linux in the future!