feat: cache metadata images on server
Retrom now caches all remote image files on
the server, and the client will fetch those
copies instead of accessing the external URLs
they were sourced from.

If a file is not currently cached, the client
will still use the upstream URL -- but the file
will be cached for next time.

fixes #366